requirements.txt
PyPDF2
langchain
langchain_google_genai
faiss-cpu
python-dotenv
google-generativeai
streamlit